Initial setup for checkout endpoint
